custom-elements/custom-element-lifecycle/types-of-callbacks/attribute-changed-callback-set-attribute-test.html

<!DOCTYPE html>
<html>
<head>
<title>Test attributeChanged callback is called if custom element attribute value is set</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<meta name="assert" content="attributeChanged callback ... must be enqueued whenever custom element's attribute is added, changed or removed.">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#types-of-callbacks">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;
    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function() {
        callbackCalledCounter++;
    }
    var GeneratedConstructor = doc.registerElement('x-a-1', {prototype: proto});
    var customElement = new GeneratedConstructor();
    customElement.setAttribute('class', 'someClass');
    assert_equals(callbackCalledCounter, 1, 'Callback attributeChanged should be called');
}, 'Test attributeChanged callback is called if attribute value is set by method ' +
    'setAttribute(). The custom element is created via constructor.');


test(function() {
    var doc = newHTMLDocument();
    var callbackArgs = null;
    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function(arg1, arg2, arg3) {
        callbackArgs = [arg1, arg2, arg3];
    }
    var GeneratedConstructor = doc.registerElement('x-b-1', {prototype: proto});
    var customElement = new GeneratedConstructor();
    customElement.setAttribute('class', 'someClass');
    assert_array_equals(callbackArgs, ['class', null, 'someClass'],
        'Unexpected callback attributeChanged arguments');
}, 'Test attributeChanged callback arguments if attribute value is set by method ' +
    'setAttribute(). The custom element is created via constructor.');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;
    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function() {
        callbackCalledCounter++;
    }
    var GeneratedConstructor = doc.registerElement('x-a-1-1', {prototype: proto});
    var customElement = new GeneratedConstructor();
    customElement.classList.add('someClass');
    assert_equals(callbackCalledCounter, 1, 'Callback attributeChanged should be called');
}, 'Test attributeChanged callback is called if attribute value is set by method ' +
    'classList.add(). The custom element is created via constructor.');


test(function() {
    var doc = newHTMLDocument();
    var callbackArgs = null;
    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function(arg1, arg2, arg3) {
        callbackArgs = [arg1, arg2, arg3];
    }
    var GeneratedConstructor = doc.registerElement('x-b-1-1', {prototype: proto});
    var customElement = new GeneratedConstructor();
    customElement.classList.add('someClass');
    assert_array_equals(callbackArgs, ['class', null, 'someClass'],
        'Unexpected callback attributeChanged arguments');
}, 'Test attributeChanged callback arguments if attribute value is set by method ' +
    'classList.add(). The custom element is created via constructor.');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;
    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function() {
        callbackCalledCounter++;
    }
    var GeneratedConstructor = doc.registerElement('x-a-1-2', {prototype: proto});
    var customElement = new GeneratedConstructor();
    customElement.className = 'someClass';
    assert_equals(callbackCalledCounter, 1, 'Callback attributeChanged should be called');
}, 'Test attributeChanged callback is called if attribute value is set as property. ' +
    'The custom element is created via constructor.');


test(function() {
    var doc = newHTMLDocument();
    var callbackArgs = null;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function(arg1, arg2, arg3) {
        callbackArgs = [arg1, arg2, arg3];
    }
    var GeneratedConstructor = doc.registerElement('x-b-1-2', {prototype: proto});
    var customElement = new GeneratedConstructor();
    customElement.className = 'someClass';

    assert_array_equals(callbackArgs, ['class', null, 'someClass'],
        'Unexpected callback attributeChanged arguments');
}, 'Test attributeChanged callback arguments if attribute value is set as property. '+
    'The custom element is created via constructor.');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function() {
        callbackCalledCounter++;
    }
    doc.registerElement('x-a-2', {prototype: proto});
    doc.body.innerHTML = '<x-a-2 id="xa2" class="theClass"></x-a-2>';

    assert_equals(callbackCalledCounter, 0, 'Callback attributeChanged should not be called');
}, 'Test attributeChanged callback is not called if attribute value is specified in HTML. '+
    'The custom element is created via innerHTML property.');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function() {
        callbackCalledCounter++;
    }
    doc.registerElement('x-a-3', {prototype: proto});
    doc.body.innerHTML = '<x-a-3 id="xa3" class="theClass"></x-a-3>';
    var customElement = doc.querySelector('#xa3');
    customElement.setAttribute('class', 'someClass');

    assert_equals(callbackCalledCounter, 1, 'Callback attributeChanged should be called');
}, 'Test attributeChanged callback is called if attribute value is set by method setAttribute(). '+
    'The custom element is created via innerHTML property.');


test(function() {
    var doc = newHTMLDocument();
    var callbackArgs = null;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function(arg1, arg2, arg3) {
        callbackArgs = [arg1, arg2, arg3];
    }
    doc.registerElement('x-b-3', {prototype: proto});
    doc.body.innerHTML = '<x-b-3 id="xb3"></x-b-3>';
    var customElement = doc.querySelector('#xb3');
    customElement.setAttribute('class', 'someClass');

    assert_array_equals(callbackArgs, ['class', null, 'someClass'],
        'Unexpected callback attributeChanged arguments');
}, 'Test attributeChanged callback arguments if attribute value is set by method setAttribute(). '+
    'The custom element is created via innerHTML property.');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function() {
        callbackCalledCounter++;
    }
    doc.body.innerHTML = '<x-a-4 id="xa4"></x-a-4>';
    var customElement = doc.querySelector('#xa4');
    // this call shouldn't invoke or enquee the callback
    customElement.setAttribute('class', 'someClass');
    assert_equals(callbackCalledCounter, 0, 'Callback attributeChanged should not be called');
    doc.registerElement('x-a-4', {prototype: proto});
    // this one should
    customElement.setAttribute('class', 'someClass2');
    assert_equals(callbackCalledCounter, 1, 'Callback attributeChanged should be called');

}, 'Test attributeChanged callback is not called if attribute value of unresolved element '+
    'is set by method setAttribute().');


test(function() {
    var doc = newHTMLDocument();
    HTML5_ELEMENTS.forEach(function(element) {
        var callbackCalledCounter = 0;
        var obj = doc.createElement(element);
        var proto = Object.create(obj.constructor.prototype);
        proto.attributeChangedCallback = function() {
            callbackCalledCounter++;
        }
        var GeneratedConstructor = doc.registerElement('x-' + element + '-' + element + '-1', {
            prototype: proto,
            extends: element
        });
        var customElement = new GeneratedConstructor();
        customElement.setAttribute('class', 'someClass');

        assert_equals(callbackCalledCounter, 1, 'Callback attributeChanged should be called');
    });
}, 'Test attributeChanged callback of the custom element that extends some HTML element. ' +
    'Test that the callback is called');


test(function() {
    var doc = newHTMLDocument();
    HTML5_ELEMENTS.forEach(function(element) {
        var callbackCalledCounter = 0;
        var obj = doc.createElement(element);
        var proto = Object.create(obj.constructor.prototype);
        proto.attributeChangedCallback = function() {
            callbackCalledCounter++;
        }
        var GeneratedConstructor = doc.registerElement('x-' + element + '-' + element + '-1-1', {
            prototype: proto,
            extends: element
        });
        var customElement = new GeneratedConstructor();
        customElement.classList.add('someClass');

        assert_equals(callbackCalledCounter, 1, 'Callback attributeChanged should be called');
    });
}, 'Test attributeChanged callback is called if attribute is set by method classList.add(). '+
    'The custom element extends some HTML element.');


test(function() {
    var doc = newHTMLDocument();
    HTML5_ELEMENTS.forEach(function(element) {
        var callbackArgs = null;

        var obj = doc.createElement(element);
        var proto = Object.create(obj.constructor.prototype);
        proto.attributeChangedCallback = function(arg1, arg2, arg3) {
            callbackArgs = [arg1, arg2, arg3];
        }
        var GeneratedConstructor = doc.registerElement('x-' + element + '-' + element + '-2', {
            prototype: proto,
            extends: element
        });
        var customElement = new GeneratedConstructor();
        customElement.setAttribute('class', 'someClass');

        assert_array_equals(callbackArgs, ['class', null, 'someClass'],
            'Unexpected callback attributeChanged arguments');
    });
}, 'Test attributeChanged callback arguments if attribute value is set by method setAttribute(). '+
    'The custom element extends some HTML element.');


testInIFrame('../../resources/x-element.html', function(doc) {
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function() {
        callbackCalledCounter++;
    }
    doc.registerElement('x-element', {prototype: proto});
    var customElement = doc.querySelector('#xelement');
    customElement.setAttribute('class', 'someClass');

    assert_equals(callbackCalledCounter, 1, 'Callback attributeChanged should be called');
}, 'Test attributeChanged callback is called if attribute value is set by method setAttribute(). '+
    'The document has browsing context.');


testInIFrame('../../resources/x-element.html', function(doc) {
    var callbackArgs = null;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function(arg1, arg2, arg3) {
        callbackArgs = [arg1, arg2, arg3];
    }
    doc.registerElement('x-element', {prototype: proto});
    var customElement = doc.querySelector('#xelement');
    customElement.setAttribute('class', 'someClass');

    assert_array_equals(callbackArgs, ['class', null, 'someClass'],
        'Unexpected callback attributeChanged arguments');
}, 'Test attributeChanged callback arguments if attribute value is set by method setAttribute(). '+
    'The document has browsing context.');


testInIFrame('../../resources/x-element.html', function(doc) {
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function() {
        callbackCalledCounter++;
    }
    doc.registerElement('x-element', {prototype: proto});
    var customElement = doc.querySelector('#xelement');
    customElement.classList.add('someClass');

    assert_equals(callbackCalledCounter, 1, 'Callback attributeChanged should be called');
}, 'Test attributeChanged callback is called if attribute value is set by method classList.add(). '+
    'The document has browsing context.');


testInIFrame('../../resources/x-element.html', function(doc) {
    var callbackArgs = null;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function(arg1, arg2, arg3) {
        callbackArgs = [arg1, arg2, arg3];
    }
    doc.registerElement('x-element', {prototype: proto});
    var customElement = doc.querySelector('#xelement');
    customElement.classList.add('someClass');

    assert_array_equals(callbackArgs, ['class', null, 'someClass'],
        'Unexpected callback attributeChanged arguments');
}, 'Test attributeChanged callback arguments if attribute value is set by method classList.add(). '+
    'The document has browsing context.');


testInIFrame('../../resources/x-element.html', function(doc) {
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attributeChangedCallback = function() {
        callbackCalledCounter++;
    }
    var customElement = doc.querySelector('#xelement');
    // this call shouldn't invoke or enquee the callback
    customElement.setAttribute('name', 'someName');
    assert_equals(callbackCalledCounter, 0, 'Callback attributeChanged should not be called');
    doc.registerElement('x-element', {prototype: proto});
    // this one should
    customElement.setAttribute('class', 'someClass');
    assert_equals(callbackCalledCounter, 1, 'Callback attributeChanged should be called');
}, 'Test attributeChanged callback is not called if attribute value of unresolved element '+
    'is set by method setAttribute(). The document has browsing context.');


testInIFrame('../../resources/x-element.html', function(doc) {
    HTML5_ELEMENTS.forEach(function(element) {
        var callbackCalledCounter = 0;
        var obj = doc.createElement(element);
        var proto = Object.create(obj.constructor.prototype);
        proto.attributeChangedCallback = function() {
            callbackCalledCounter++;
        }
        var GeneratedConstructor = doc.registerElement('x-' + element + '-' + element + '-1', {
            prototype: proto,
            extends: element
        });
        var customElement = new GeneratedConstructor();
        customElement.setAttribute('class', 'someClass');

        assert_equals(callbackCalledCounter, 1, 'Callback attributeChanged should be called');
    });
}, 'Test attributeChanged callback is called if attribute value is set by method setAttribute(). '+
    'The document has browsing context. The custom element extends some HTML element.');


testInIFrame('../../resources/x-element.html', function(doc) {
    HTML5_ELEMENTS.forEach(function(element) {
        var callbackArgs = null;

        var obj = doc.createElement(element);
        var proto = Object.create(obj.constructor.prototype);
        proto.attributeChangedCallback = function(arg1, arg2, arg3) {
            callbackArgs = [arg1, arg2, arg3];
        }
        var GeneratedConstructor = doc.registerElement('x-' + element + '-' + element + '-2', {
            prototype: proto,
            extends: element
        });
        var customElement = new GeneratedConstructor();
        customElement.setAttribute('class', 'someClass');

        assert_array_equals(callbackArgs, ['class', null, 'someClass'],
            'Unexpected callback attributeChanged arguments');
    });
}, 'Test attributeChanged callback arguments if attribute value is set by method setAttribute(). '+
    'The document has browsing context. The custom element extends some HTML element.');
</script>
</body>
</html>
